==> 6-behaviorColored_Trace.ipynb <==
import os
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt


import warnings
warnings.filterwarnings("ignore")

# prep info
# contact: c
# status inlcude: r, s, a

root_path = r'/Users/xiaoqiansun/Desktop/MedLu/TubeTest/Data'
fig_path = r'/Users/xiaoqiansun/Desktop/MedLu/TubeTest/Trace'


days = 6
fpsN = 15
fpsB = 30
baselinePeriod = 180 #(3min(180s) baseline)
redundentTime = 190 #10s+3min


M32_list = ['32-1_28', '32-2_18', '32-3_15', '32-4_20', '32-5_12', '32-6_15']
M33_list = ['33-1_28', '33-2_35', '33-3_32', '33-4_30', '33-5_30', '33-6_26']



# Define Some Functions

def plot_allNeuronTrace_withColoredBehavior(df,folderPath,which_mouse,mouse_name,fig_path,decovolute=False):

    df1 = df.copy().dropna().reset_index(drop=True)
    
    dfs = df1[df1[which_mouse]== 's']
    dfr = df1[df1[which_mouse]== 'r']
    dfa = df1[df1[which_mouse]== 'a']
    dfc = df1[df1['contact']== 'c']
    
    
    x = df1.index.values.tolist()
    xs = dfs.index.values.tolist()
    xr = dfr.index.values.tolist()
    xa = dfa.index.values.tolist()
    xc = dfc.index.values.tolist()

    
    all_neurons = df1.drop(['mouse1', 'mouse2', 'contact'], axis=1).columns.tolist()
    # read in ON/OFF summary dataframe
    file_name = 'ONOFF.csv'
    on_off_df = pd.read_csv(os.path.join(folderPath, 'ONOFF.csv'))
    columns = on_off_df.columns.to_list()
    columns[0] = 'Neurons'
    on_off_df.columns = columns
    on_off_df = on_off_df.set_index('Neurons')
    aON = on_off_df[on_off_df['approachON'] == 1].index.to_list()
    rON = on_off_df[on_off_df['retreatON'] == 1].index.to_list()
        
        
    fig, ax = plt.subplots(len(all_neurons),1, figsize=(40,3*len(all_neurons)),facecolor='w', edgecolor='k')

    for i in range(len(all_neurons)):
        neuronName = all_neurons[i]

        y = df_frame_label.iloc[:,i].values.tolist()
        if decovolute:
            c, s, b, g, lam = deconvolve(np.asarray(y), penalty=1)
            y=b+c

        ax[i].plot(x, y, lw=4, color = 'black')

        for xcs in xs:
            ax[i].axvline(x=xcs, color='#4f4f4f', alpha=0.2) #grey, still
        for xcr in xr:
            ax[i].axvline(x=xcr, color='#7c83ff', alpha=0.2) # blue, retreat
        for xca in xa:
            ax[i].axvline(x=xca, color='#FF7C80', alpha=0.2) # pink, approach
        for xcc in xc:
            ax[i].axvline(x=xcc, color='#f8ff7c', alpha=0.2) # yellow, concat

        if neuronName in aON:
            ax[i].set_title(neuronName+" approach ON", fontsize=28, fontweight='bold')
        elif neuronName in rON:
            ax[i].set_title(neuronName+" retreat ON", fontsize=28, fontweight='bold')
        else:
            ax[i].set_title(neuronName, fontsize=28, fontweight='bold')

        ax[i].spines['right'].set_visible(False)
        ax[i].spines['top'].set_visible(False)
        ax[i].spines['bottom'].set_visible(False)
        ax[i].spines['left'].set_visible(False)
        ax[i].axes.get_xaxis().set_visible(False)
        ax[i].axes.get_yaxis().set_visible(False)

    
    fig.savefig(os.path.join(fig_path, mouse_name+'_allNeuronTrace.png'))
    
    plt.close()

# Call

for day in range(days):
    
    # read in dataframe
    m32_folder = os.path.join(root_path, M32_list[day])
    m33_folder = os.path.join(root_path, M33_list[day])
    
    
    m32_tube = pd.read_csv(os.path.join(m32_folder,'m32_tube.csv'), index_col = 'Frame')
    m33_tube = pd.read_csv(os.path.join(m33_folder,'m33_tube.csv'), index_col = 'Frame')    
    
    # plot neuron trace
    #----------------------------------------------------------------------------------
    plot_allNeuronTrace_withColoredBehavior(m32_tube,m32_folder,'mouse1',
                                            'mouse32_'+str(day+1),fig_pathdecovolute=False)
    plot_allNeuronTrace_withColoredBehavior(m33_tube,m33_folder,'mouse2',
                                            'mouse33_'+str(day+1),fig_pathdecovolute=False)


    break
